refactor(*): split promisePolyfill and clearAnimations from index.js
